+++ Module.php
@@ -2,7 +2,6 @@
 
 namespace humhub\modules\cfiles;
 
-use humhub\components\console\Application as ConsoleApplication;
 use humhub\modules\cfiles\models\ConfigureContainerForm;
 use humhub\modules\cfiles\models\rows\FileSystemItemRow;
 use humhub\modules\space\models\Space;
@@ -17,6 +16,16 @@ use yii\helpers\Url;
 class Module extends ContentContainerModule
 {
     /**
+     * @var int Files uploaded into the same folder by the same user are announced by a single
+     *      notification, once no further file was uploaded for this many minutes.
+     *      0 announces every upload request on its own.
+     *
+     * @see \humhub\modules\cfiles\libs\FileUploadBatch
+     * @since 0.19
+     */
+    public int $uploadNotificationDelay = 10;
+
+    /**
      * @var string sort name as 'name', 'size', 'updated_at'
      * @see FileSystemItemRow::ORDER_MAPPING
      */
@@ -29,19 +38,6 @@ class Module extends ContentContainerModule
      */
     public $defaultPostedFilesSort = FileSystemItemRow::ORDER_TYPE_UPDATED_AT;
     public $defaultPostedFilesOrder = SORT_ASC;
-
-    /**
-     * @inheritdoc
-     */
-    public function init()
-    {
-        parent::init();
-
-        if (Yii::$app instanceof ConsoleApplication) {
-            // Prevents the Yii HelpCommand from crawling all web controllers and possibly throwing errors at REST endpoints if the REST module is not available.
-            $this->controllerNamespace = 'cfiles/commands';
-        }
-    }
 
     /**
      * @inheritdoc