+++ messages/bg/base.php
@@ -116,4 +116,7 @@ return [
     'Wrong target folder!' => '',
     'You cannot move the file "{name}"!' => '',
     'You cannot move the folder "{name}"!' => '',
+    '{displayName} added {n,plural,=1{a file} other{# files}} to the folder "{folderTitle}".' => '',
+    '{originator} added {n,plural,=1{a file} other{# files}} to the folder "{folderTitle}"' => '',
+    '{originator} added {n,plural,=1{a file} other{# files}} to the folder "{folderTitle}" in Space {space}' => '',
 ];