+++ messages/pt/base.php
@@ -116,4 +116,7 @@ return [
     'Folder content ID' => '',
     'Folders' => '',
     'Hide in Stream' => '',
+    '{displayName} added {n,plural,=1{a file} other{# files}} to the folder "{folderTitle}".' => '',
+    '{originator} added {n,plural,=1{a file} other{# files}} to the folder "{folderTitle}"' => '',
+    '{originator} added {n,plural,=1{a file} other{# files}} to the folder "{folderTitle}" in Space {space}' => '',
 ];